+++ so_long/includes/get_next_line.h
@@ -22,7 +22,7 @@
 
 typedef struct s_fopen
 {
-	int		read_size;
+	ssize_t	read_size;
 	char	*buff;
 	size_t	str_size;
 	int		is_newline;
